feat(crafter.go): add implementation details
in this change we are introducing the initial ranking of recipes by profit to the FindProfitableOptions method.
It is debatable if we want to only work with already profitable options or not. Also we still need to think about how to handle vendor itens and intermediate contents for recipes.

BREAKING CHANGE: FindProfitableOptions return type has changed so it has to be properly handled now.